Fixed main docopt setup and use of logging
The logging facility was used in a wrong way by passing
the log object around. This commit fixes this part and
also puts the docopt string to the python top scope were
it belongs to. The result of this code is now also covered
by tests which makes sure to handle errors the way we
want it