Refactor generator
Turn the generator into a class. Apply the single responsibilty
OO pattern. Add tests for the code and give the single parts
of the processing a name. Update the keg commandline code to
make use of the classes. Keg should also be usuable as an API,
thus I think it's important come up with a good class design.
This Fixes #10