Preserve XML comments
The KIWI validation does not preserve comments after validation
as they have no meaning for the processing of the image description.
However for OBS comments are treated as project configurations
and therefore the toplevel comments if present must be added back.

Because comment processing is tricky especially for inline comments
and when moving to a another than the XML markup, this commit also
includes a processing refactoring. In addition to the toplevel
comment preserve on XML documents it also adds format processing
options as follows:

Added two new options --format-xml|--format-yaml which allows
to explicitly specify if on top of the Keg generated KIWI
description an update to the current KIWI schema should be
applied or not. In addition a warning about handling of
comments in this mode will be issued

If no format-* options are given the keg generated XML document
is only validated against the KIWI API but not translated into
the installed schema. This allows to keep all comments as
created by the keg jinja template mechanism.

This Fixes #19