Add support for images.sh script
Unfortunately adding support for another script hook was not
possible without changing the layout of the image definition
data for scripts. The basic structure for the scripts config.sh
and images.sh is now as follows:

```yaml
config:
  # config.sh setup
  config_script:
    NAMESPACE:
      - SCRIPTLET_NAME
    files:
      NAMESPACE:
        - PATH_DEFINITION
    services:
      NAMESPACE:
        - SERVICE_DEFINITION
    sysconfig:
      NAMESPACE:
        - FILE_DEFINITION

  # images.sh setup
  image_script:
    SAME_STRUCTURE_AS_FOR_CONFIG_SCRIPT
```

On merge of this commit the data structure in the keg-recipes
repository must be adapted. This Fixes #35