script template file is not related to schema type
The script template file names for config.sh.templ, images.sh.templ
were prefixed by the schema type, e.g 'vm'. This is not needed
because the contents of the scripts are defined by the collected
data from the image definition tree and do not change in ther
semantic depending on the schema type. Therefore we only need
one template for config.sh and one template for images.sh