add unit test coverage for delete-key support
Adds a key to a (not actually produced) Leap <= 15.1 image description and
removes it for the Leap 15.2 one that is used in the unit test. This will cover
the corresponding code path.

This also required a slight rearrangment of the mockup image defintion, as it
wasn't quite using the appropriate keg data layout.