OBS source service: add support for change log generator
Add support for updating image change logs using keg's change log generator.
Changes are inserted at the beginning for an existing (flavor.)changes.yaml
file and namespaced with the image version.
The latest commit's hash from each repository is stored in _keg_revisions
(which needs to be stored in the OBS package) and will be used in subsequent
runs to select the next version's starting point.
This also changes the version increment implementation to not use kiwi
to write a new config.kiwi file, as this removes all comments.