image schema fixes
Allow empty 'profiles' section in image description. This allows to declare an
empty section in a stub image definition so the place where the profiles
section can be defined but still be used with single-build definitions that
do not have a profiles section.
Add missing namespace string definition for 'services'.
Change 'overlay' to '_include_overlays' in 'archive' section.