XML generator: fix list type parameter handling
When converting a dict to param=string pairs, allow list values to indicate
that the given param is used multiple times. E.g. 'console: [ttyS0, tty1]'
would be converted to 'console=ttyS0 console=tty1'. An empty list is considered
a flag param, i.e. no '=string' part. This was supported before but also
considered a value of 0 to be a flag, which was not intended and is fixed.